backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
prisma==0.15.0
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
slowapi==0.1.9
structlog==24.4.0
python-multipart==0.0.17
httpx==0.28.1
